[perm] fixes for user permissions
