+++ frappe/core/page/permission_manager/permission_manager.py
@@ -7,7 +7,7 @@ import frappe.defaults
 from frappe.modules.import_file import get_file_path, read_doc_from_file
 from frappe.translate import send_translations
 from frappe.desk.notifications import delete_notification_count_for
-from frappe.permissions import reset_perms
+from frappe.permissions import reset_perms, get_linked_doctypes
 
 @frappe.whitelist()
 def get_roles_and_doctypes():
@@ -30,15 +30,6 @@ def get_permissions(doctype=None, role=None):
 		(doctype and (" parent='%s'" % doctype.replace("'", "\'")) or "",
 		role and ((doctype and " and " or "") + " role='%s'" % role.replace("'", "\'")) or ""),
 		as_dict=True)
-
-	def get_linked_doctypes(dt):
-		return list(set([dt] + [d.options for d in
-			frappe.get_meta(dt).get("fields", {
-				"fieldtype":"Link",
-				"ignore_user_permissions":("!=", 1),
-				"options": ("!=", "[Select]")
-			})
-		]))
 
 	linked_doctypes = {}
 	for d in out:
